updated if statements

+++ general/kvm/buildvyos.sh
@@ -7,6 +7,12 @@ fi
 
 randomid=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 4 | head -n 1)
 servername=test-$randomid
+iso=/mnt/data/isos/vyos-1.1.7-amd64.iso
+
+if [ ! -f $iso ]; then
+	echo "File $iso not found"
+	exit
+fi
 
 virt-install -n $servername --ram=512 --vcpu=1 --disk pool=datastore_1,format=qcow2,size=5 \
 	--graphics vnc,listen=0.0.0.0,password=password --noautoconsole \
